requirements.txt
aiohttp
aiosignal
async-timeout
attrs
Brotli
certifi
cffi
charset-normalizer
click
colorama
cryptography
frozenlist
idna
joblib
multidict
mutagen
nltk
openai
prompt-toolkit
pycparser
pycryptodomex
pypng
pyshorteners
python-dotenv
qrcode
regex
requests
tqdm
typing_extensions
urllib3
wcwidth
websockets
yarl
yt-dlp